Mobile progress, screens for meditation and food created, auto login and logout

+++ Assignments/year3/mobile/mobile-assignments/app/src/main/java/com/example/assignment1/Models/Entry.kt
@@ -7,8 +7,8 @@ import java.time.LocalDateTime
 @Entity(tableName = "entry")
 data class Entry(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0, // Unique identifier for each entry
-    val userId: Int?, // Id attached to user
+    val id: Long = 0, // Unique identifier for each entry
+    val userId: Long?, // Id attached to user
     var text: String? = null, // Text entered by the user
     val name: String?, // User who entered the text
     var date: LocalDateTime, // Date of entry
